+++ tensorflow_quantum/core/ops/tfq_simulate_state_op_cuquantum.cu.cc
@@ -145,7 +145,7 @@ class TfqSimulateStateOpCuQuantum : public tensorflow::OpKernel {
     // Simulate programs one by one. Parallelizing over state vectors
     // we no longer parallelize over circuits. Each time we encounter a
     // a larger circuit we will grow the Statevector as necessary.
-    for (int i = 0; i < fused_circuits.size(); i++) {
+    for (size_t i = 0; i < fused_circuits.size(); i++) {
       int nq = num_qubits[i];
 
       if (nq > largest_nq) {
@@ -155,7 +155,7 @@ class TfqSimulateStateOpCuQuantum : public tensorflow::OpKernel {
         sv_host.resize(2 * (uint64_t(1) << largest_nq));
       }
       ss.SetStateZero(sv);
-      for (int j = 0; j < fused_circuits[i].size(); j++) {
+      for (size_t j = 0; j < fused_circuits[i].size(); j++) {
         qsim::ApplyFusedGate(sim, fused_circuits[i][j], sv);
       }
